Revert "lists: mark wget-nossl as unwanted alternative"
There are some packages, which has dependency on wget.
Like foris-subordinates-plugin, but it installs wget-nossl.
Ref. https://forum.turris.cz/t/turris-os-4-0-2-is-released-into-hbt-testing-branch/11428/33?u=pepe

It is possible that it needs to have different priority, but let's solve
it once for all. Don't build wget-nossl. We don't need to save a few KBs.

This reverts commit 748caaf5cb6c278ed54d762884515084017d0cf8.