configs: clean and rework
There are various changes done to configuration on top of just general
cleanup.

All constants from compile_pkgs were moved to configs as it makes sense
that they are in one location not in script.

The settings to use specific version of toolchain was removed. Now we
use upstream default toolchain version which prevents problems with some
packages not being compiled because of old toolchain is used.

All options settings support and capabilities of target were dropped as
they are overriden by target setting non the less so they make no sense.
These should be modified in OpenWrt target not in configuration.

All package settings were dropped with exception of essential packages
and package disablements.
The reason why essential packages were not dropped is to ensure that all
failed build won't be deployed. With future tests done as part of build
this file can be dropped.
And lastly we have to preserve disablement of packages to not build
them as intended.

All DEFAULT options were dropped as they are setting default for symbol
and are intended to be set from target but all of them are overruled by
setting that all packages should be build (CONFIG_ALL).