configs/common/kernel: drop unused options
These options are not in use. They are not activated because of missing
dependency activation. Both of them require module to be enabled and
that module has to packaged. In case of NFT the packages are already in
upstream. In case of XT packages are missing and should possibly be
created in upstream.